+++ src/protocol/mod.rs
@@ -1,2 +1,12 @@
 pub mod commands;
 pub mod events;
+
+/// Size of the key for a file node.
+pub const FILE_KEY_SIZE: usize = 32;
+/// Size of the key for a file node.
+pub const FOLDER_KEY_SIZE: usize = 16;
+
+/// Size of the user's key.
+pub const USER_KEY_SIZE: usize = 16;
+/// Size of the user's sid.
+pub const USER_SID_SIZE: usize = 43;