+++ CHANGELOG.md
@@ -13,11 +13,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added `Client::resume_session` method.
 - Added `Client::serialize_session` method.
+- Added `Client::fetch_protected_nodes` method.
 
 ### Changed
 
 - Added `#[non_exhaustive]` attribute on `Error` enum.
 - Changed `Error` variants from tuples to named fields.
+- Changed `Client` to implement the `Send` and `Sync` traits.
+- Changed `ClientBuilder::timeout` method to take an `Option<Duration>` instead of just `Duration`.
 
 ### Fixed