+++ CHANGELOG.md
@@ -21,10 +21,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed `Error` variants from tuples to named fields.
 - Changed `Client` to implement the `Send` and `Sync` traits.
 - Changed `ClientBuilder::timeout` method to take an `Option<Duration>` instead of just `Duration`.
+- Upgraded dependencies
 
 ### Fixed
 
 ### Removed
+
+- Removed `Error::Pbkdf2Error` enum variant.
 
 [0.5.0] - 2023-06-08
 --------------------